Split up tests into multiple, logically split files
