+++ Tests/Helpers.swift
@@ -1,0 +1,21 @@
+//
+//  Helpers.swift
+//  PusherSwift
+//
+//  Created by Hamilton Chapman on 07/04/2016.
+//
+//
+
+import Foundation
+
+func convertStringToDictionary(text: String) -> [String:AnyObject]? {
+    if let data = text.dataUsingEncoding(NSUTF8StringEncoding) {
+        do {
+            let json = try NSJSONSerialization.JSONObjectWithData(data, options: .MutableContainers) as? [String:AnyObject]
+            return json
+        } catch {
+            print("Something went wrong")
+        }
+    }
+    return nil
+}
